backend/requirements.txt
asgiref==3.11.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
waitress==2.1.2
